Remove auto_type and auto_source from simple EDXML writer
Now that event types and sources are no longer attributes
of event groups it makes sense to move the responsibility
of assigning the type and source URI of events to the
instantiator of the events. The EDXML writer will now
require all events that it receives for output to have
an event type and source URI set.