Improve XML tree cleanup code
The code to clean up the in memory XML tree while transcoding
XML data was rather naive. We used to delete every element
visited by the parser regardless of it being trancoded or not.
Also, we used to delete all preceding sibling nodes in stead
of specifically deleting the previously transcoded element.