Simplify tuning event type definitions in transcoders
We split off a new class method that generates a single event type
from the class attributes of the transcoder. This makes it simpler
for transcoders to override this method and tune the event type
definitions. Also, it allows calling the method to generate an
event type without instantiating the transcoder.